[roll third_party/pigweed] pw_allocator: Remove total_bytes metric
Previously, `pw::allocator::Metrics` had a field for `total_bytes`. This
is not particularly interesting, as it is a value that isn't expected
to change. It also is a bit difficult, as a `TrackingAllocator` has no
way to interrogate an `Allocator` it is handed for its implementation-
specific capacity. As a result, this CL removes the field.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/193251

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: a5038affded6fea..e8c65f4cc0c1568
Roller-URL: https://ci.chromium.org/b/8755359665830692225
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: If66ae621cd2a73b4f7303b59e4f20972e206678b
Reviewed-on: https://pigweed-review.googlesource.com/c/example/echo/+/193617
Bot-Commit: Pigweed Roller <pigweed-roller@pigweed-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Pigweed Roller <pigweed-roller@pigweed-service-accounts.iam.gserviceaccount.com>